requirements.txt
streamlit==1.38.0
pandas==2.2.2
altair==5.4.1
langchain==0.3.7
langchain-openai==0.2.8
langchain-core==0.3.15
openai==1.45.0
tiktoken==0.7.0
python-dotenv==1.0.1
Pillow==10.4.0
plotly==5.24.0
folium==0.17.0
requests==2.32.3
